refactor(clickhouse): extract staging store setup into staging.go
Move createStagingStore, createS3StagingStore, and createGCSStagingStore
into a dedicated file to keep the clickhouse.go diff minimal. Remove
unused ValidateS3 and ValidateStaging functions.
